Add the IAccessibleTextSelectionContainer interface which allows getting and setting text selections within descendant content. (#17)
IAccessible2 currently provides no easy, consistent way of getting and setting selections within a document or other content container.
The IAccessibleText interface does allow for getting, adding and removing selections, however this is only really useful for plain text controls.
IAccessible2_3 allows retrieval of selections across objects. However, there was a lot of ambiguity in what the offsets meant, and the caller was not forced to use IAccessibleText objects.

This PR adds an IAccessibleTextSelectionContainer interface with a selections getter and a setSelections method. Selections are represented by an IA2TextSelection struct. This is different to the struct in IAccessible2_3:

* The selection ranges are made up of start and end object/offset pairs, rather than anchor and active object/offset pairs.
* The selection ranges now have a `startIsActive` boolean.
* The start and end objects must support the IAccessibleText interface.
* The start and end offset are always character offsets into the IAccessibleText interface, rather than child offsets such as in w3c DOM selections.

This deprecates the IAccessible2_3 interface.

Signed-off-by: James Teh <jteh@mozilla.com>