Convert text/object attribute spec to Markdown
Drop the table of contents that has links to
the corresponding sections in the Linux Foundation
wiki.

One thing noticeable when comparing the wiki page
and the converted Markdown version (or an HTML version
generated from that) is that the text

> The prefix is inserted in front of the text. The prefix contains
> a number (with any leading or trailing punctuation such as leading
> and/or trailing parens and/or a trailing period), or a bullet
> character, or the word “graphic”.

for the "prefix" list subattribute in the object attributes spec
does not show up on the Linux Foundation wiki page [1]
while it does here.
This text also shows up when looking at the pagesource [2]
in the wiki, though, so having it seems to be appropriate.

Fixes: #24

[1] https://wiki.linuxfoundation.org/accessibility/iaccessible2/objectattributes
[2] https://wiki.linuxfoundation.org/accessibility/iaccessible2/objectattributes?do=edit

Signed-off-by: Michael Weghorn <m.weghorn@posteo.de>